services/streaming/requirements.txt
pyspark==3.5.2
kafka-python==2.0.2
httpx==0.27.2
pydantic==2.9.2
psycopg[binary]==3.2.3
pydantic-settings==2.5.2
